+++ lib/cpp/services/point.hpp
@@ -4,7 +4,7 @@ namespace ngPlasmid
 {
     struct Point
     {
-        long double x;
-        long double y;
+        float x;
+        float y;
     };
 }
